+++ eslint.config.mjs
@@ -2,7 +2,7 @@ import antfu from '@antfu/eslint-config'
 
 export default antfu(
   {
-    ignores: ['dist', 'info', 'ui/.angular'],
+    ignores: ['dist'],
     rules: {
       'curly': ['error', 'all'],
       'no-undef': 'error',
@@ -41,3 +41,31 @@ export default antfu(
     },
   },
 )
+  .append({
+    files: ['src/**/*.test.ts'],
+    rules: {
+      'test/expect-expect': 'error',
+      'test/no-commented-out-tests': 'error',
+      'test/no-conditional-expect': 'error',
+      'test/no-disabled-tests': 'warn',
+      'test/no-focused-tests': 'error',
+      'test/no-identical-title': 'error',
+      'test/no-import-node-test': 'error',
+      'test/no-interpolation-in-snapshots': 'error',
+      'test/no-mocks-import': 'error',
+      'test/no-standalone-expect': 'error',
+      'test/no-unneeded-async-expect-function': 'error',
+      'test/padding-around-before-each-blocks': 'error',
+      'test/padding-around-after-each-blocks': 'error',
+      'test/padding-around-before-all-blocks': 'error',
+      'test/padding-around-after-all-blocks': 'error',
+      'test/padding-around-describe-blocks': 'error',
+      'test/padding-around-test-blocks': 'error',
+      'test/prefer-called-exactly-once-with': 'error',
+      'test/require-local-test-context-for-concurrent-snapshots': 'error',
+      'test/valid-describe-callback': 'error',
+      'test/valid-expect': 'error',
+      'test/valid-expect-in-promise': 'error',
+      'test/valid-title': 'error',
+    },
+  })